requirements.txt
numpy==1.23.2
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
streamlit==1.12.2
tensorflow==2.9.1
